up czi update

+++ workflows/czi_predict.py
@@ -9,8 +9,8 @@
 #######################################################################
 
 import asyncio
-import shutil
-from os.path import join
+from os import mkdir
+from os.path import join, isdir
 
 import boto3
 import click
@@ -38,14 +38,15 @@ from tardis_em.utils.device import get_device
 from tardis_em.utils.predictor import Predictor
 from tardis_em.utils.normalization import RescaleNormalize, MeanStdNormalize
 from tardis_em.utils.export_data import to_mrc
-from shutil import rmtree
+from shutil import rmtree, move
 
-def get_from_aws(url):
+
+async def get_from_aws(url):
     s3 = boto3.client("s3", config=Config(signature_version=UNSIGNED))
     s3.download_file("cryoet-data-portal-public", url, url.split("/")[-1])
 
 
-def upload_to_aws(aws_key, aws_secret, bucket, id_, data_name, file_name):
+async def upload_to_aws(aws_key, aws_secret, bucket, id_, data_name, file_name, remove=True):
     client = boto3.client(
         "s3", aws_access_key_id=aws_key, aws_secret_access_key=aws_secret
     )
@@ -55,15 +56,23 @@ def upload_to_aws(aws_key, aws_secret, bucket, id_, data_name, file_name):
         bucket,
         join("robert_kiewisz_tardis_01_2024", id_, data_name, file_name),
     )
-    rmtree(file_name)
+    if remove:
+        rmtree(file_name)
+    else:
+        move(file_name, join('TARDIS', file_name))
 
 
 class ProcessTardisForCZI:
-    def __init__(self, allocate_gpu: int):
+    def __init__(self, allocate_gpu: int, predict='Membrane'):
         self.device = get_device(allocate_gpu)
-        self.normalize_px = 15
+        self.predict = predict
+        self.normalize_px = 15 if self.predict == "Membrane" else 25
         self.image_stitcher = StitchImages()
         self.post_processes = BuildPointCloud()
+
+        self.output = join("temp", "Predictions")
+        self.patches = join("temp", "Patches")
+        self.results = "TARDIS"
 
         # Normalize histogram
         self.normalize = RescaleNormalize(clip_range=(1, 99))
@@ -74,7 +83,7 @@ class ProcessTardisForCZI:
             checkpoint=None,
             network="unet",
             subtype="32",
-            model_type="membrane_3d",
+            model_type="membrane_3d" if self.predict == "Membrane" else "microtubules_3d",
             img_size=128,
             sigmoid=False,
             device=self.device,
@@ -84,9 +93,34 @@ class ProcessTardisForCZI:
             checkpoint=None,
             network="dist",
             subtype="triang",
-            model_type="3d",
+            model_type="3d" if self.predict == "Membrane" else "2d",
             device=self.device,
         )
+
+        self.patch_pc = PatchDataSet(max_number_of_points=1000, graph=False)
+        self.segment_pc = PropGreedyGraphCut(threshold=0.5,
+                                             connection=8 if self.predict == "Membrane" else 2)
+
+    @staticmethod
+    def clean_temp():
+        if isdir("temp"):
+            rmtree("temp")
+
+    def build_temp(self):
+        """
+        Standard set-up for creating new temp dir for cnn prediction.
+
+        Args:
+            dir_ (str): Directory where folder will be build.
+        """
+        if not isdir(self.results):
+            mkdir(self.results)
+
+        self.clean_temp()
+
+        mkdir("temp")
+        mkdir(self.patches)
+        mkdir(self.output)
 
     def predict_cnn(self, dataloader):
         for j in range(len(dataloader)):
@@ -94,11 +128,28 @@ class ProcessTardisForCZI:
             input_ = self.model_cnn.predict(input_[None, :], rotate=True)
             tif.imwrite(join(self.output, f"{name}.tif"), input_)
 
-    def __call__(self, data: np.ndarray, px: float, header, name):
-        data = self.normalize(self.mean_std(data)).astype(np.float32)
-        tif.imwrite(name + ".tif", data.sum(0).astype(np.uint16))
+    def predict_dist(self, coords_df):
+        x = []
+        with torch.no_grad():
+            for i in coords_df:
+                x.append(
+                    self.model_dist.predict(x=i[None, :].to(self.device), y=None)
+                    .cpu()
+                    .detach()
+                    .numpy()[0, 0, :]
+                )
 
-        # Sanity check image histogram
+        return x
+
+    def __call__(self, data: np.ndarray, px: float, header, name):
+        """Build temp dir"""
+        self.build_temp()
+
+        """Load and prepare dataset for prediction"""
+        data = self.normalize(self.mean_std(data)).astype(np.float32)
+        tif.imwrite(join(self.results, name + ".tif"),
+                    data.sum(0).astype(np.float32))
+
         if not data.min() >= -1 or not data.max() <= 1:  # Image not between in -1 and 1
             if data.min() >= 0 and data.max() <= 1:
                 data = (data - 0.5) * 2  # shift to -1 - 1
@@ -116,49 +167,41 @@ class ProcessTardisForCZI:
             scale=scale_shape,
             trim_size_xy=128,
             trim_size_z=128,
-            output=join("temp", "Patches"),
+            output=self.patches,
             image_counter=0,
             clean_empty=False,
             stride=round(128 * 0.125),
         )
 
+        """TARDIS - Semantic Segmentation"""
         self.predict_cnn(
-            dataloader=PredictionDataset(join("temp", "Patches", "imgs")),
+            dataloader=PredictionDataset(join(self.patches, "imgs")),
         )
         data = self.image_stitcher(
-            image_dir=join("temp", "Predictions"), mask=False, dtype=np.float32
+            image_dir=self.output, mask=False, dtype=np.float32
         )[: scale_shape[0], : scale_shape[1], : scale_shape[2]]
-        data, _ = scale_image(image=data, scale=org_shape, nn=True)
+        data, _ = scale_image(image=data, scale=org_shape, nn=False)
         data = torch.sigmoid(torch.from_numpy(data)).cpu().detach().numpy()
 
         data = np.where(data > 0.25, 1, 0).astype(np.uint8)
-        to_mrc(data, px, name + "_semantic.mrc", header)
-        tif.imwrite(name + "_semantic.tif", data.sum(0).astype(np.uint8))
+        to_mrc(data, px, name + "_semantic.mrc", header)  # To upload back to AWS
+        tif.imwrite(join(self.results, name + "_semantic.tif"),
+                    data.sum(0).astype(np.uint8))  # Store for NYSBC
 
+        """TARDIS - Instance Segmentation"""
         _, pc_ld = BuildPointCloud().build_point_cloud(
             image=data, EDT=False, down_sampling=5, as_2d=False
         )
-
-        coords_df, _, output_idx, _ = PatchDataSet(
-            max_number_of_points=900, graph=False
-        ).patched_dataset(coord=pc_ld)
-        x = []
-        with torch.no_grad():
-            for i in coords_df:
-                x.append(
-                    self.model_cnn.predict(x=i[None, :].to(self.device), y=None)
-                    .cpu()
-                    .detach()
-                    .numpy()[0, 0, :]
-                )
-
-        pc = PropGreedyGraphCut(threshold=0.5, connection=8).patch_to_segment(
+        coords_df, _, output_idx, _ = self.patch_pc.patched_dataset(coord=pc_ld)
+        x = self.predict_dist(coords_df)
+        pc = self.segment_pc.patch_to_segment(
             graph=x,
             coord=pc_ld,
             idx=output_idx,
-            sort=False,
-            prune=15,
+            sort=False if self.predict == "Membrane" else True,
+            prune=15 if self.predict == "Membrane" else 5,
         )
+
         pc = pd.DataFrame(pc)
         pc.to_csv(
             name + "_instance.csv",
@@ -194,21 +237,21 @@ async def main(aws_dir: str, bucket: str, allocate_gpu: int):
     terminal(title='Predict CIZ-Cryo-EM data-port datasets')
     process_tardis = ProcessTardisForCZI(allocate_gpu)
 
-    aws = np.genfromtxt(aws_dir, delimiter=",")
+    # AWS pre-setting
+    aws = np.genfromtxt(aws_dir, delimiter=",", dtype=str)
     aws_key = aws[0]
     aws_secret = aws[1]
-
     client = Client()
 
     # Get all tomograms with voxel spacing <= 10
-    tomos = list(
+    all_tomogram = list(
         Tomogram.find(
             client,
         )
     )
 
     # S3 URIs for MRCs
-    urls = [t.s3_mrc_scale0 for t in tomos]
+    urls = [t.s3_mrc_scale0 for t in all_tomogram]
     folder_names = [s[31:36] for s in urls]
     file_names = [s.split("/")[4] for s in urls]
     czi_px = [float(s.split("/")[6][12:]) for s in urls]
@@ -224,9 +267,9 @@ async def main(aws_dir: str, bucket: str, allocate_gpu: int):
             zip(urls, folder_names, file_names, czi_px)
     ):
         terminal(title='Predict CIZ-Cryo-EM data-port datasets',
-                 text_1='Prediction:',
+                 text_1='Waiting for Data from AWS:',
                  text_3=f'Dataset: [{file_name}]',
-                 text_5=print_progress_bar(idx, len(urls)), )
+                 text_6=print_progress_bar(idx, len(urls)), )
 
         # Wait for the current download to complete before processing
         await next_download
@@ -237,6 +280,12 @@ async def main(aws_dir: str, bucket: str, allocate_gpu: int):
         if px != px_czi:
             px = px_czi
 
+        terminal(title='Predict CIZ-Cryo-EM data-port datasets',
+                 text_1='TARDIS segmentation:',
+                 text_3=f'Dataset: [{file_name}]',
+                 text_4=f'Pixel_size: {px}',
+                 text_6=print_progress_bar(idx, len(urls)), )
+
         process_task = asyncio.create_task(
             asyncio.get_running_loop().run_in_executor(
                 None, process_tardis, data, px, header, file_name[:-4]
@@ -244,7 +293,7 @@ async def main(aws_dir: str, bucket: str, allocate_gpu: int):
         )
 
         # If there's a next file, start downloading it now
-        if url != urls[-1]:
+        if idx > len(file_names):
             next_download = asyncio.create_task(get_from_aws(url[31:]))
 
         await process_task
@@ -269,6 +318,7 @@ async def main(aws_dir: str, bucket: str, allocate_gpu: int):
                     folder_names,
                     file_name[:-4],
                     file_name[:-4] + "_instance.csv",
+                    remove=False,
                 )
             ),
         ]
